Extract `pictype` to own module
