+++ src/catacomb.rs
@@ -11,7 +11,7 @@ use crate::{
     catasm::{cgarefresh, drawchartile, egarefresh},
     class_type::classtype::{self, *},
     control_struct::ControlStruct,
-    cpanel::{controlpanel, installgrfile, pictype, spritetype},
+    cpanel::{controlpanel, installgrfile, spritetype},
     cpanel_state::CpanelState,
     demo_enum::demoenum,
     dir_type::dirtype::{self, *},
@@ -38,6 +38,7 @@ use crate::{
         SDL_DisplayMode, SDL_GameController, SDL_Rect,
     },
     pcrlib_c_state::PcrlibCState,
+    pic_type::pictype,
     rleasm::RLEExpand,
     safe_sdl::{SDL_Renderer, SDL_Texture, SDL_Window, SDL_mutex},
     scan_codes::*,
