+++ src/pic_type.rs
@@ -1,0 +1,8 @@
+#[derive(Copy, Clone)]
+#[repr(C, packed)]
+pub struct pictype {
+    pub width: i16,
+    pub height: i16,
+    pub shapeptr: u32,
+    pub name: [i8; 8],
+}
